requirements.txt
# requirements.txt
flask
gunicorn
flask-sqlalchemy
requests
python-dateutil
python-dotenv